# Conflicts:
#	core/local/src/main/java/com/club/local/CoreDataStoreDataSourceImp.kt
#	core/repositories/src/main/java/com/devfalah/repositories/CoreDataStoreDataSource.kt

+++ identity/useCases/src/main/java/com/thechance/identity/usecases/SaveUserIdUseCase.kt
@@ -6,6 +6,6 @@ class SaveUserIdUseCase @Inject constructor(
     private val identityRepository: IdentityRepository
 ) {
     suspend operator fun invoke(id: String){
-        identityRepository.saveUserId(id)
+        identityRepository.saveUserId(id.toInt())
     }
 }
